♻️ JE-36-refactor: content 내부 컴포넌트 리팩토링
