+++ src/components/ActivityDetail/DetailContent/Title.tsx
